fix bug report #1806540, report & fix by John Evans (gnugeek)

+++ classes/database/Postgres.php
@@ -4,7 +4,7 @@
  * A class that implements the DB interface for Postgres
  * Note: This class uses ADODB and returns RecordSets.
  *
- * $Id: Postgres.php,v 1.305 2007/09/29 09:09:45 ioguix Exp $
+ * $Id: Postgres.php,v 1.306 2007/10/02 21:36:35 ioguix Exp $
  */
 
 // @@@ THOUGHT: What about inherits? ie. use of ONLY???
@@ -644,7 +644,7 @@ class Postgres extends ADODB_base {
 		$sql = "SELECT NULL AS nspname, c.relname, 
 					(SELECT usename FROM pg_user u WHERE u.usesysid=c.relowner) AS relowner, 
 					(SELECT description FROM pg_description pd WHERE c.oid=pd.objoid) AS relcomment,
-					reltuples::integer AS reltuples
+					reltuples::bigint AS reltuples
 				FROM pg_class c 
 				WHERE c.relkind='r' 
 					AND NOT EXISTS (SELECT 1 FROM pg_rewrite r WHERE r.ev_class = c.oid AND r.ev_type = '1')
